Add BaseElasticsearchDocumentStore
