+++ haystack/document_stores/__init__.py
@@ -11,18 +11,20 @@ from haystack.document_stores.es_converter import elasticsearch_index_to_documen
 from haystack.document_stores.es_converter import open_search_index_to_document_store
 
 try:
+    # Use appropriate ElasticsearchDocumentStore depending on ES client version
     with LazyImport() as elasticsearch_import:
-        # Use appropriate ElasticsearchDocumentStore depending on ES client version
         from elasticsearch import __version__ as ES_VERSION
 
         if ES_VERSION[0] == 7:
             from haystack.document_stores.elasticsearch7 import ElasticsearchDocumentStore  # type: ignore  # pylint: disable=reimported,ungrouped-imports
         elif ES_VERSION[0] == 8:
-            if not typing.TYPE_CHECKING:
-                from haystack.document_stores.elasticsearch8 import ElasticsearchDocumentStore  # type: ignore  # pylint: disable=reimported,ungrouped-imports
+            from haystack.document_stores.elasticsearch8 import ElasticsearchDocumentStore  # type: ignore  # pylint: disable=reimported,ungrouped-imports
+        else:
+            # Use Elasticsearch 7 as default
+            from haystack.document_stores.elasticsearch7 import ElasticsearchDocumentStore
     elasticsearch_import.check()
 except (ModuleNotFoundError, ImportError):
-    # Use Elasticsearch 7 as default
+    # No need to import anything if ES could not be imported
     from haystack.document_stores.elasticsearch7 import ElasticsearchDocumentStore
 
 from haystack.document_stores.opensearch import OpenSearchDocumentStore
